# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/add": {
            "post": {
                "description": "行内以下描述",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "用户相关"
                ],
                "summary": "添加用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "pswd",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/del/{name}": {
            "delete": {
                "description": "行内以下描述",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户相关"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "提交的参数",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/find/{name}": {
            "get": {
                "description": "根据name查询用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户相关"
                ],
                "summary": "查询用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "要查询的名字",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/regis": {
            "post": {
                "description": "行内以下描述",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户相关"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "提交的参数",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/test": {
            "get": {
                "description": "行内以下描述",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "测试"
                ],
                "summary": "测试",
                "responses": {}
            }
        },
        "/upd/{name}": {
            "patch": {
                "description": "根据name更新用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户相关"
                ],
                "summary": "更新用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "被更改的名字",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更改的参数",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "models.User": {
            "type": "object",
            "properties": {
                "mid": {
                    "description": "消息id",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "portrait": {
                    "description": "头像地址",
                    "type": "string"
                },
                "pswd": {
                    "description": "密码",
                    "type": "string"
                },
                "stateMsg": {
                    "description": "线上状态消息",
                    "type": "string"
                }
            }
        }
    }
}